requirements.txt
# Django Backend Dependencies
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
python-dotenv>=1.0

# Production Server
gunicorn>=21.0

# Database (Production)
psycopg2-binary>=2.9

# AWS S3 Storage (Optional - for media files)
boto3>=1.34
django-storages>=1.14
